requirements.txt
systemd-python==234
opencv-python==4.5.5.62